Created lang module and removed unescessary type decorating
