+++ src/js/rishson/control/ControllerWidget.js
@@ -27,18 +27,6 @@ define([
 
 		/**
 		 * @field
-		 * @name rishson.control.ControllerWidget.types
-		 * @type {Object}
-		 * @description a collection of the different widget types, used to decorate the objects
-		 */
-		types: {
-			MODEL: "model",
-			CONTROLLER: "controller",
-			VIEW: "view"
-		},
-
-		/**
-		 * @field
 		 * @name rishson.control.ControllerWidget.views
 		 * @type {Object}
 		 * @description a collection of all the injected views
@@ -57,8 +45,6 @@ define([
 		 * @constructor
 		 */
 		constructor: function () {
-			this.type = this.types.CONTROLLER;
-
 			this.models = {};
 			this.views = {};
 			this.controllers = {};
@@ -79,8 +65,6 @@ define([
 		 * @param {Object|rishson.widget._Widget} view a widget
 		 */
 		addView : function (view, name) {
-			view.type = this.types.VIEW;
-
 			this.views[name] = view;
 		},
 
@@ -93,8 +77,6 @@ define([
 		 * @description create placeholder for a model and register handler for listener callbacks.
 		 */
 		addModel : function (name, model, topicName) {
-			model.type = this.types.MODEL;
-
 			var myModel = this.models[name] = model;	//lookup shortcut
 			myModel.listeners = [];
 			myModel.loaded = false;
