+++ src/js/rishson/control/_Controller.js
@@ -4,8 +4,9 @@ define([
 	"rishson/Globals", //TOPIC_NAMESPACE
 	"dojo/_base/array", // forEach
 	"dojo/_base/lang", // hitch
+	"rishson/base/lang",
 	"dojo/topic" // publish, subscribe
-], function (declare, Base, Globals, arrayUtil, lang, topic) {
+], function (declare, Base, Globals, arrayUtil, lang, rishsonLang, topic) {
 	/**
 	 * @class
 	 * @name rishson.control._Controller
@@ -66,6 +67,32 @@ define([
 					this._wireSinglePub(child.pubList[topicObj]);
 				}
 			}
+		},
+
+		/**
+		 * @function
+		 * @name rishson.control._Controller._unAutoWireControllerPubs
+		 * @description Un-subscribes this controller from subscriptions to the supplied widget
+		 * @param {Object} The widget containing a list of published items
+		 */
+		_unAutoWirePubs: function (widget) {
+			var pubList = widget.pubList || (widget.content || {}).pubList, // We want the actual widget if this widget is a ContentPane
+				topicNamespace = this._topicNamespace;
+
+			// Loop through the child widgets pubList
+			rishsonLang.forEachObjProperty(pubList, function (pubHandleName) {
+				// If the current controllers namespace appears within this widgets pubList item
+				if (pubHandleName.indexOf(topicNamespace) !== -1) {
+					var handle = this.subListHandles[pubHandleName];
+
+					// If a handle was found then remove the subscription
+					if (handle) {
+						console.log("Unsubscribed from: " + pubHandleName);
+						this.unsubscribe(handle);
+						delete this.subListHandles[pubHandleName];
+					}
+				}
+			}, this);
 		},
 
 		/**
